Move SnapshotHierarchy to snapshots package
