+++ subprojects/snapshots/src/test/groovy/org/gradle/internal/snapshot/SnapshotUtilTest.groovy
@@ -16,7 +16,6 @@
 
 package org.gradle.internal.snapshot
 
-import org.gradle.internal.vfs.SnapshotHierarchy
 import spock.lang.Specification
 import spock.lang.Unroll
